refactoring: PathFinder의 findPath 구현 리펙토링

+++ src/test/java/nextstep/subway/path/acceptance/PathSteps.java
@@ -6,7 +6,7 @@ import io.restassured.response.Response;
 
 public class PathSteps {
 
-    public static ExtractableResponse<Response> 경로_조회_요청(int source, int target) {
+    public static ExtractableResponse<Response> 경로_조회_요청(Long source, Long target) {
         return RestAssured
                 .given().log().all()
                 .params("source", source)
